added sliders
