+++ src/app/globals.scss
@@ -9,6 +9,9 @@
 
 html,
 body {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
   max-width: 100vw;
   overflow-x: hidden;
 }
